TEST-002: Restructure test_auth_database.py under 500-line limit
Move TestAPIKeyCRUD (7 tests) from test_auth_database.py to
test_auth_api_keys.py where it joins the high-level API key lifecycle
tests. Remove temp file-exemptions.yml entry.

- test_auth_database.py: 560 → 473 lines
- test_auth_api_keys.py: 198 → 278 lines
- All 848 unit tests pass
