+++ docs/file-exemptions.yml
@@ -120,12 +120,6 @@ exemptions:
 # Removed: dango/platform/network.py — now 486 lines, within limit (TASK-088)
 
   # --- Test files (verbose by nature — each test class covers one public function) ---
-  - file: tests/unit/test_auth_database.py
-    lines: 560
-    category: exempt
-    reason: "At limit from TASK-010; TASK-014 adds update_api_key_last_used test. TEST-002 will split into multiple files."
-    review_by: "v1.0"
-
   - file: tests/unit/test_oauth_validation.py
     lines: 580
     category: exempt
